% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCJ02_WGS84.R
\name{GCJ02_WGS84}
\alias{GCJ02_WGS84}
\title{GCJ02_WGS84}
\usage{
GCJ02_WGS84(lon, lat)
}
\arguments{
\item{lon}{original longitude based on the GCJ02 coordinate system.}

\item{lat}{original latitude based on the GCJ02 coordinate system.}
}
\value{
longtitude and latitude based on the WGS84 coordinate system.
}
\description{
GCJ02_WGS84
}
\examples{
GCJ02_WGS84(100.00,40.00)
}
