% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode.cityStateZip.create}
\alias{geocode.cityStateZip.create}
\title{geocodes.cityStateZip.create}
\usage{
geocode.cityStateZip.create(x)
}
\arguments{
\item{x}{google api address results}
}
\description{
creates cityStateZip from google query
}
\keyword{api,}
\keyword{cityStateZip}
\keyword{google,}
