\name{geocode.google}
\alias{geocode.google}
\title{Returns the longitude and latitude of a geocoded point using the Google
Maps API.}
\usage{
  geocode.google(address, key = "anonymous")
}
\arguments{
  \item{address}{the street address to geocode.}

  \item{key}{the Google Maps API key.}
}
\value{
  a geocoded point.
}
\description{
  Returns the longitude and latitude of a geocoded point
  using the Google Maps API.
}

