% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode.street.create}
\alias{geocode.street.create}
\title{geocodes.street.create}
\usage{
geocode.street.create(x)
}
\arguments{
\item{x}{google api address results}
}
\description{
creates street from google query
}
\keyword{api,}
\keyword{google,}
\keyword{street}
