% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{geocode_api}
\alias{geocode_api}
\title{Geocoding with Google API}
\usage{
geocode_api(u_address, return.call = "json")
}
\arguments{
\item{u_address}{desired location which will return results.}

\item{return.call}{the api coding format json or xml.}
}
\value{
returns print output of full address and latitude , longitude.
}
\description{
This function returns the location details.
}
\examples{
geocode_api("dhaka")

}
\references{
\href{https://developers.google.com/maps/documentation/geocoding/intro}{GoogleGeoAPI}
}
