% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_amap.R
\name{Get_amap}
\alias{Get_amap}
\title{Get_amap}
\usage{
Get_amap(address, city = NULL, key = "156a263e434219631f61d4e2e402f0cb")
}
\arguments{
\item{address}{detailed address in Chinese}

\item{city}{which city does it belong to}

\item{key}{amap API key: '156a263e434219631f61d4e2e402f0cb' or 'e885b568bdb9500b97cf1f46d572ff1e'}
}
\value{
formatted_address, country, province, city, district, longitude, latitude based on GCJ02/WGS84 coordinate system.
}
\description{
Get_amap
}
\examples{
Get_amap('北京大学')
}
