% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGS84_GCJ02.R
\name{WGS84_GCJ02}
\alias{WGS84_GCJ02}
\title{WGS84_GCJ02}
\usage{
WGS84_GCJ02(lng, lat)
}
\arguments{
\item{lat}{original latitude based on the WGS84 coordinate system.}

\item{lon}{original longitude based on the WGS84 coordinate system.}
}
\value{
longitude and latitude based on the GCJ02 coordinate system.
}
\description{
WGS84_GCJ02
}
\examples{
WGS84_GCJ02(100.00,40.00)
}
