% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geocode.R
\docType{package}
\name{Geocode}
\alias{Geocode}
\alias{Geocode-package}
\title{A package connecting to a web API.}
\description{
The lab contains creation of a package which will connect to google geocode API
and gives us the desired locations full address, latitude and longitude.
}
\section{geocode_api function}{

geocode_api() returns the desired locations full address, latitude and longitude.
}

\examples{
geocode_api("lund")
}
\references{
\href{https://developers.google.com/maps/documentation/geocoding/intro}{GoogleGeoAPI}
}
