\name{geocode.yahoo}
\alias{geocode.yahoo}
\title{Returns the longitude and latitude of a geocoded point using the Yahoo
Maps API.}
\usage{
  geocode.yahoo(address, key)
}
\arguments{
  \item{address}{the street address to geocode.}

  \item{key}{the Yahoo Maps API key.}
}
\value{
  a geocoded point.
}
\description{
  Returns the longitude and latitude of a geocoded point
  using the Yahoo Maps API.
}
\references{
  \url{http://developer.yahoo.com/geo/placefinder/guide/}
}

