% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocodes.bad.address}
\alias{geocodes.bad.address}
\title{geocodes.bad.address}
\usage{
geocodes.bad.address(l.pkg.path, geocodes.bad.address.new, location.source)
}
\arguments{
\item{l.pkg.path}{character vector l.pkg.path(geocodes.bad.address.location, parcels)}

\item{geocodes.bad.address.new}{new bad geocoded address}

\item{location.source}{location source}
}
\description{
Creates and loads points.address file
}
\keyword{clean,}
\keyword{geocode}
\keyword{points,}
